requirements.txt
PyOpenGL
PyOpenGL_accelerate
glfw==2.6.1
numpy